move worker to azure exporter, will refactor in the next PR
