+++ opencensus/common/schedule/__init__.py
@@ -136,37 +136,3 @@ class Queue(object):
         else:
             for item in items:
                 self.put(item, block, timeout)
-
-
-class Worker(threading.Thread):
-    daemon = True
-
-    def __init__(self, src, dst):
-        self.src = src
-        self.dst = dst
-        self._stopping = False
-        super(Worker, self).__init__()
-
-    def run(self):
-        src = self.src
-        dst = self.dst
-        while True:
-            batch = src.gets(dst.max_batch_size, dst.export_interval)
-            if batch and isinstance(batch[-1], QueueEvent):
-                dst.emit(batch[:-1], event=batch[-1])
-                if batch[-1] is src.EXIT_EVENT:
-                    break
-                else:
-                    continue
-            dst.emit(batch)
-
-    def stop(self, timeout=None):
-        start_time = time.time()
-        wait_time = timeout
-        if self.is_alive() and not self._stopping:
-            self._stopping = True
-            self.src.put(self.src.EXIT_EVENT, block=True, timeout=wait_time)
-            elapsed_time = time.time() - start_time
-            wait_time = timeout and max(timeout - elapsed_time, 0)
-        if self.src.EXIT_EVENT.wait(timeout=wait_time):
-            return time.time() - start_time  # time taken to stop
